requirements.txt
# # Standard libraries
# python-dotenv
# datetime
# # Datasets and libraries
# datasets
# transformers
# requests
# scikit-learn
# Pillow
# torchvision
# pandas
# # Models and pipelines
# peft
# bitsandbytes
# accelerate

# # Model tokenization
# sentencepiece
# tqdm




# Core ML
torch>=2.0.0
torchvision
torchaudio

# Transformers & HuggingFace
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0
tokenizers>=0.15.0

# Evaluation & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0
evaluate>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# NLP Tools
spacy>=3.7.0
nltk>=3.8.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.16.0

# Utilities
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# For LLaMA/Mistral access
huggingface-hub>=0.19.0

# For 4-bit quantization support
# auto-gptq>=0.5.0
optimum>=1.16.0